backend/requirements.txt
fastapi[all]
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
openai
pydantic[email]